% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annar.R
\name{roll}
\alias{roll}
\title{Dice}
\usage{
roll(dietype = 1:6, probvalues = rep(1/length(dietype), length(dietype)))
}
\arguments{
\item{dietype}{The type of die.}

\item{probvalues}{The probability of getting a particular number of divots.}
}
\value{
The sum of two dice.
}
\description{
Dice
}
\examples{
set.seed(123)
roll()
roll(dietype = 1:20)
}

