% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annar.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{A value chosen from a deck.}
}
\value{
Random cards being chosen showing face, suit, and value.
}
\description{
Shuffle
}
\examples{
deck2[1:6, ]
deck2[2,3]
}

